// src/common/enums/dates.ts
export enum Dates {
  TODAY = "today",
  YESTERDAY = "yesterday",
  TWO_DAYS_AGO = "two_days_ago",
  LAST_WEEK = "last_week",
  WEEK_CUSTOM = "custom_week",
  CUSTOM = "custom",
  CURRENT_WEEK = "current_week",
  MONTH_CUSTOM = "custom_month",
  CURRENT_MONTH = "current_month",
  LAST_MONTH = "last_month",
}

export enum DateTexts {
  today = "Today's",
  yesterday = "Yesterday's",
  two_days_ago = "2 days ago",
  current_week = "Current week's",
  last_week = "Last week's",
  current_month = "Current month's",
  last_month = "Last month's",
}